refactor: implementation of browser tab title
