+++ ui/index.html
@@ -9,7 +9,7 @@
       name="viewport"
     />
     <meta content="noindex,nofollow" name="robots" />
-    <title>Halo</title>
+    <title></title>
     <link rel="icon" href="/favicon.ico" type="image/png" />
     <link rel="icon" href="/favicon.ico" type="image/svg+xml" />
     <meta name="msapplication-TileColor" content="#da532c" />
